moved input parsing to parser module
