+++ index.js
@@ -30,22 +30,7 @@ console.log(`Style: ${program.style}`);
 console.log(`Watch: ${program.watch}`);
 
 
-let entries = [];
-if (fs.lstatSync(path.resolve(process.cwd(), input)).isDirectory()) {
-    fs.readdir(input, (err, files) => {
-        files.forEach(file => {
-            let fullname = path.resolve(process.cwd(), input, file);
-            if (!fs.lstatSync(fullname).isDirectory()) {
-                entries.push(fullname);
-            }
-        });
-        sortEntries();
-    });
-} else {
-    entries = fg.sync([input], { dot: true });
-}
-
-function sortEntries() {
+parser.parseInput(input, function (entries) {
     console.log(`Entries: ${entries}`);
 
     entries.forEach(function (filename) {
@@ -53,7 +38,7 @@ function sortEntries() {
             renderSheet(filename);
         }
     });
-}
+});
 
 function renderSheet(filename) {
     console.log(`Rendering ${filename}`)
